refactor: move AssetPath into a separate file for use in validator
