+++ src/aind_data_schema/components/configs.py
@@ -24,7 +24,8 @@ from pydantic_core.core_schema import ValidationInfo
 from aind_data_schema.base import DataModel, GenericModelType
 from aind_data_schema.components.coordinates import Coordinate, CoordinateSystem, Scale, Transform
 from aind_data_schema.components.tile import AcquisitionTile, Channel
-from aind_data_schema.components.identifiers import Code, AssetPath
+from aind_data_schema.components.identifiers import Code
+from aind_data_schema.components.wrappers import AssetPath
 
 
 class StimulusModality(str, Enum):
