+++ src/aind_data_schema/core/processing.py
@@ -8,7 +8,8 @@ from aind_data_schema_models.units import MemoryUnit, UnitlessUnit
 from pydantic import Field, SkipValidation, ValidationInfo, field_validator, model_validator
 
 from aind_data_schema.base import AwareDatetimeWithDefault, DataCoreModel, DataModel, GenericModel, GenericModelType
-from aind_data_schema.components.identifiers import Code, Person, AssetPath
+from aind_data_schema.components.identifiers import Code, Person
+from aind_data_schema.components.wrappers import AssetPath
 from aind_data_schema.utils.merge import merge_notes
