+++ src/aind_data_schema/utils/validators.py
@@ -2,8 +2,10 @@
 
 from enum import Enum
 from typing import Any, List, Optional
-from pathlib import Path, PurePosixPath
+from pathlib import Path
 import logging
+
+from aind_data_schema.components.wrappers import AssetPath
 
 # Fields that should have the same length as the coordinate system axes
 AXIS_FIELD_NAMES = ["scale", "translation", "angles", "position"]
@@ -117,13 +119,13 @@ def recursive_check_paths(obj: Any, directory: Optional[Path] = None):
     if isinstance(obj, Enum):
         return
 
-    if isinstance(obj, PurePosixPath):
-        obj = Path(obj)
-
-    if isinstance(obj, Path):
+    if isinstance(obj, AssetPath):
         full_path = directory / obj if directory else obj
+        full_path = Path(full_path)
         if not full_path.exists():
-            logging.warning(f"AssetPath {full_path} does not exist")
+            logging.warning(
+                f"AssetPath {full_path} does not exist, ensure file paths are relative to the metadata directory"
+            )
     elif isinstance(obj, (list, tuple, set, dict)):
         items = obj.values() if isinstance(obj, dict) else obj
         for item in items:
